capacitación: flujo completo por skills + cadena de encoders por hardware
Skillifica el flujo de punta a punta para usarlo como eje de una capacitación:

- /empezar — puerta de entrada por proyecto (trigger estricto, solo el comando):
  crea video-N, recibe el crudo, transcribe UNA vez (preview de costo si es nube)
  y bifurca con la pregunta ¿editar para redes o convertir a shorts?
- /broll-ai — del transcript al video con b-roll generado (estética promptimagen
  embebida), con DOBLE gate: momentos+prompts+costo antes de generar imágenes, y
  aprobación imagen por imagen antes de animar. Blotato principal, fal.ai respaldo.
  tools/insert_broll.py inserta desde broll-plan.json con verificación automática.
- /subtitulos — quemado determinístico desde edited-transcript.json vía
  tools/make_subs.py (.ass + .srt, gate de frame de muestra, estilo desde brand.md,
  audio copiado bit a bit).
- /publicar — cierre vía Blotato: lista las redes conectadas EN VIVO, copy por
  plataforma, publicar/agendar, gate duro de confirmación. Marcador
  PENDIENTE-LIGA-VIDEO para el tutorial de conexión.
- /ayuda — mentor del flujo: entrega la imagen guía (media/library/guia/), detecta
  en qué paso va el proyecto y responde dudas desde las fuentes del repo.
- tools/hwenc.py — cadena de encoders NVENC → Quick Sync → AMF → CPU con prueba
  funcional por juego completo de args (preview y final por separado) y pistas de
  driver viejo; render_cuts.py y make_proxy.py la consumen (make_proxy tenía NVENC
  cableado sin fallback).
- pricing.json §broll — política de costos de generación (créditos consultados en
  vivo, nunca precios inventados).
- format_transcript.py — clock() acepta ms float (fix menor preexistente).

+++ tools/format_transcript.py
@@ -20,8 +20,8 @@ def fmt_time(ms: int) -> str:
     return f"{ms / 1000:.2f}"
 
 
-def clock(ms: int) -> str:
-    s = ms // 1000
+def clock(ms: float) -> str:
+    s = int(ms) // 1000  # acepta ms float (p.ej. audio_duration del backend local)
     return f"{s // 60:02d}:{s % 60:02d}"
